// src/types.ts
export type PdfToJsonOptions = {
    input: string;
    output: string;
    config: string;
    license: string;
};

export type PdfToXlsxOptions = {
    input: string;
    output: string;
    license: string;
};

export type XlsxToJsonOptions = {
    input: string;
    output: string;
    config: string;
};